char_dim introduced
char_dim as separate from model_dim. Add FC layer to upsample to model_dim. That FC layer is now effectively the embedding layer, but its inputs are enriched by character embs.

This hopefully saves compute in the char self-attn; all the linear layers there have to be called 16 times more often than in the token-self-attn layers, because the sequence length is so much longer.